+++ psl-core/src/main/java/org/linqs/psl/application/inference/online/messages/actions/controls/Sync.java
@@ -17,14 +17,14 @@
  */
 package org.linqs.psl.application.inference.online.messages.actions.controls;
 
-import org.linqs.psl.application.inference.online.messages.actions.OnlineAction;
+import org.linqs.psl.application.inference.online.messages.OnlineMessage;
 
 /**
  * Sync the atom values of the online model.
  * Specifically, execute queued actions and optimize.
  * String format: Sync
  */
-public class Sync extends OnlineAction {
+public class Sync extends OnlineMessage {
     public Sync() {
         super();
     }
