+++ psl-core/src/main/java/org/linqs/psl/application/inference/online/messages/actions/template/ActivateRule.java
@@ -15,23 +15,23 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.linqs.psl.application.inference.online.messages.actions.template.actions;
+package org.linqs.psl.application.inference.online.messages.actions.template;
 
 import org.linqs.psl.model.rule.Rule;
 
 /**
- * Add a new rule to the model.
- * String format: AddRule <READ/WRITE> <predicate> <args> ... <value>
+ * Activate an existing rule in the model.
+ * String format: ActivateRule <rule>
  */
-public class AddRule extends TemplateAction {
-    public AddRule(Rule rule) {
+public class ActivateRule extends TemplateAction {
+    public ActivateRule(Rule rule) {
         super(rule);
     }
 
     @Override
     public String toString() {
         return String.format(
-                "ADDRULE\t%s",
+                "ACTIVATERULE\t%s",
                 rule.toString());
     }
 }
