+++ psl-core/src/main/java/org/linqs/psl/application/inference/online/messages/actions/template/TemplateAction.java
@@ -15,9 +15,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.linqs.psl.application.inference.online.messages.actions.template.actions;
+package org.linqs.psl.application.inference.online.messages.actions.template;
 
-import org.linqs.psl.application.inference.online.messages.actions.OnlineAction;
+import org.linqs.psl.application.inference.online.messages.OnlineMessage;
 import org.linqs.psl.model.atom.Atom;
 import org.linqs.psl.model.predicate.ExternalFunctionalPredicate;
 import org.linqs.psl.model.rule.Rule;
@@ -25,7 +25,7 @@ import org.linqs.psl.model.rule.Rule;
 import java.util.HashSet;
 import java.util.Set;
 
-public abstract class TemplateAction extends OnlineAction {
+public abstract class TemplateAction extends OnlineMessage {
     protected Rule rule;
 
     public TemplateAction(Rule rule) {
@@ -46,8 +46,7 @@ public abstract class TemplateAction extends OnlineAction {
         return rule;
     }
 
-    public Rule setRule(Rule rule) {
+    public void setRule(Rule rule) {
         this.rule = rule;
-        return rule;
     }
 }
